Move model-related functions from wagtail.admin.utils to wagtail.admin.models
